fix: resolve all clippy warnings (25 → 0)
- Collapse 18 nested if-let chains (collapsible_if)
- Remove unnecessary parentheses (watcher.rs)
- Change &PathBuf → &Path in dir_label/update_active_label (ptr_arg)
- Replace while-let-on-iterator with for loop (terminal_state.rs)
- Suppress too_many_arguments on pixel helper functions (3 sites)
- Fix unnecessary cast (tab_bar.rs)

+++ src/watcher.rs
@@ -72,8 +72,8 @@ impl Drop for FsWatcher {
 
 // ── inotify background thread ──────────────────────────────────────
 
-const WATCH_MASK: u32 = (libc::IN_CREATE | libc::IN_DELETE | libc::IN_MODIFY
-    | libc::IN_MOVED_FROM | libc::IN_MOVED_TO) as u32;
+const WATCH_MASK: u32 = libc::IN_CREATE | libc::IN_DELETE | libc::IN_MODIFY
+    | libc::IN_MOVED_FROM | libc::IN_MOVED_TO;
 
 fn spawn_watcher(
     path: PathBuf,
